Replicate OOP logic flow from AB3
