+++ src/main/java/seedu/address/logic/commands/Command.java
@@ -2,6 +2,7 @@ package seedu.address.logic.commands;
 
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.model.Model;
+import seedu.address.model.WeddingModel;
 
 /**
  * Represents a command with hidden internal logic and the ability to be executed.
@@ -15,6 +16,19 @@ public abstract class Command {
      * @return feedback message of the operation result for display
      * @throws CommandException If an error occurs during command execution.
      */
-    public abstract CommandResult execute(Model model) throws CommandException;
+    public CommandResult execute(Model model) throws CommandException {
+        throw new UnsupportedOperationException("This command does not support execution on a Model.");
+    }
 
+    /**
+     * Executes the command on a WeddingModel.
+     *
+     * @param model {@code WeddingModel} which the command should operate on.
+     * @return feedback message of the operation result for display
+     * @throws CommandException If an error occurs during command execution.
+     */
+    public CommandResult execute(WeddingModel model) throws CommandException {
+        throw new UnsupportedOperationException("This command does not support execution on a WeddingModel.");
+    }
 }
+
