+++ src/main/java/seedu/address/logic/commands/wedding/AddWeddingPersonCommand.java
@@ -7,7 +7,7 @@ import seedu.address.logic.Messages;
 import seedu.address.logic.commands.Command;
 import seedu.address.logic.commands.CommandResult;
 import seedu.address.logic.commands.exceptions.CommandException;
-import seedu.address.model.Model;
+import seedu.address.model.WeddingModel;
 import seedu.address.model.person.Person;
 import seedu.address.model.wedding.Wedding;
 
@@ -31,7 +31,7 @@ public class AddWeddingPersonCommand extends Command {
     }
 
     @Override
-    public CommandResult execute(Model model) throws CommandException {
+    public CommandResult execute(WeddingModel model) throws CommandException {
         requireNonNull(model);
 
         if (model.weddingHasPerson(wedding, toAdd)) {
