+++ src/main/java/seedu/address/model/Model.java
@@ -6,7 +6,6 @@ import java.util.function.Predicate;
 import javafx.collections.ObservableList;
 import seedu.address.commons.core.GuiSettings;
 import seedu.address.model.person.Person;
-import seedu.address.model.wedding.Wedding;
 
 /**
  * The API of the Model component.
@@ -85,31 +84,5 @@ public interface Model {
      * @throws NullPointerException if {@code predicate} is null.
      */
     void updateFilteredPersonList(Predicate<Person> predicate);
-
-    /**
-     * Adds the given wedding.
-     * {@code Wedding} must not already exist in the Wedding Planner.
-     */
-    void addWedding(Wedding wedding);
-
-    /**
-     * Returns true if a wedding with the same identity as {@code Wedding} exists in the Wedding Planner.
-     */
-    boolean hasWedding(Wedding wedding);
-
-    /**
-     * Checks for duplicate Person in {@code wedding}.
-     * @param wedding
-     * @param toAdd
-     * @return true if the Wedding already has a Person with the same identity as {@code Person}
-     */
-    boolean weddingHasPerson(Wedding wedding, Person toAdd);
-
-    /**
-     * Adds a Person to the given Wedding.
-     * @param wedding
-     * @param toAdd
-     */
-    void addWeddingPerson(Wedding wedding, Person toAdd);
 
 }
